use merge tool of filterSuite to generate reports
This additional step can even be used to check
if the files describing the partitions contain all the
test cases. The pipeline will show a warning if this is
not the case. This makes sure all defined tests have been run
and a report has been generated. The allure report will *not*
contain any tests not defined in any of the white lists.