Fixed: [L3] deleting cookies on loginsite leads to timout on login
Root cause: UI code did not check, if indexeddb is still present or in a closing state. Therefore, these error where not catched and the UI hung up
Solution: Catch error and continue without a indexeddb. This will not cache any files for the next page load but prevent the UI from stalling.