Fixed: Bug 57323 - favorites folder missing after some time/suspend
Root cause: The favorite folders are received with a /PUT multiple request. If the network connection is shaky, this request might fail. Instead of different errors for the distinct requests inside the multiple, the response just contains a single error. This error will be used to reject every single get request. We had a mechanism, which sets the ids for an error to identify the object for which the request failed. This code part assumes, that every error is a different object. If it is not, the id of the error would have been set to the last in the list and therefore, the response looks as if there was just one folder present. All other folders were removed and just one folder remained.
Solution: Just use a copy of the error to make sure, that the id's are not overwritten.