Fixed: OXUIB-183 - [L3] Found no such composition space
Root cause: There is a little time gap between the POST /compose/:id/attachment to state to have a progress of 100% and the fact, that the upload call resolves. This is the time, the server needs to finally store the attachment somewhere. If the mail is send in exactly this gap, a race condition between sending and attaching the image to the mail might occur.
Solution: Wait until the attachment-upload has been resolved before the mail send process can be started.