WIP: C8826: Cancel adding attachment
Sadly there doesn't seem to be a way to interact with the native FileChooser to cancel it. I've tried adding a helper that uses Puppeteer directly, but that also doesn't work because the Chromium 80 and Puppeteer don't work together to expose the FileChoosers properly either. Lastly, I'm not sure how Attachments are visualized in the new mail compose so the check to see if no attachments have been added isn't part of the test yet. Maybe it's worth revisiting this branch at some point. Maybe it's all for the trash bin.