fix tests according to middleware change
there is a difference between permissions and capabilities, and the MW
team decided to make the App Suite throw an error if some API is not used
as designed.
Some provisioning has been done in an unintended way, this is corrected now
but needs an update of the codecept-helper library.