don't block pipeline with manual jobs
after switching to rules syntax for job definitions,
the pipeline gets “blocked” if a manual action is still
required. This is the case for all the cleanup jobs,
which will run automatically if the environment goes away.
We don't want those jobs to block the entire pipeline, because
MR can't be merged in that case.