Fixed OXUIB-476 Already loaded Contacts are not reloaded if they are changed on the middleware
root cause: The last modified value for the contact was already changed by an list all request,
but the values of the contact was unchanged. Additionally there was no redraw of the contact on change events.
solution: Mark the contact in cache as changed, when more than the last modified value was changed.
Also add a trigger for a redraw on change.