[DOCS-1985] rework drive new menu
This does change the naming and order of some actions in the drive toolbar.
A new 'upload' dropdown is also introduced.
On mobile the actions have a the name of the parent dropdown as a prefix, because the paranet name from the dropdown (i.e. 'New', 'Upload') is not visible.