Fix: OXUIB-627 - Appointments jump to current week on click
Root cause: The short drag started the drag and drop mechanism of the appointment. Usually, this will not be a problem, but if the appointment is not within the displayed timeframe, the drag and drop mechanism does not work.
Solution: Disable drag and drop for appointments, that are not within the visible timeframe.

(cherry picked from commit 5283461aa02224afefa1056f1a443ad2de50ada9)