[e2e] fix test for fixed-width fonts
for some reason this test reported green for a long time, but should have been
red since basically forever. A change introduced lately showed the problems of
this test, which should now be fixed. It's now waiting for the correct element
and checking for class names instead of css properties which might change more often.