OXUIB-1333: immediate HTML text formating with signature
Setting tinymce's formatting to `''` instead of `'raw'` did the
trick. Raw-mode would lead to the removal of a zero width no-break
space, which leaves the span to which the formatting would be
applied empty. Because it's empty, the span gets deleted. Changing the
formating also led to minor changes in the text's html. Therefore,
some tests needed to be adapted. Also, when removing a signture below
a blockquote left an trailing blank line, which needs removal.