OXUIB-768 Follow-up appointments: purging old conference data
When creating a follow-up for an appointment with a conference,
the old conference's data is deleted from the location and
description fields. If no data is removed from the description
field, a warning informing the user about possible outdated data
in the description is issued. This is useful, when the data might
have been edited manually and could not be deleted.

Fixed outdated conference data in follow-up appointments.
Added warning for potential outdated conference data in follow-up
appointments.