Fixed: OXUIB-1703 - mail compose jumps when insert link
Root cause: In the 4.x stream of tinymce, the paste plugin somehow causes this jump.
Solution: Add a workaround to retain the position after a paste. This might result i a short flicker, but there is no other way to fix this for 7.10.x. The better fix comes with 8.0 and the major update of tinymce to 5.x