fixed OXUIB-1795 - XSS using "upsell" triggers
Root cause: Capabilities for checking can come from dynamic sources (in this case the jslobs) and if the code-snippet doesn't contain any letters, it will just be interpreted and therefore be executed.
Solution: Presanitize the inserted code to only allow a set of whitelisted characters. Those characters are either logical connectors or are going to be replaced with true/false in a subsequent replacement call. Therefore, only logical operations but no executions are possible anymore.