fix: OXUIB-1890: Inserting text into mail body using copy/paste does not work all the time
Root cause: Event order of tinymce is inconsistent
Solution: Be robust when event order is wrong. The scrollposition must only be fixed when the order is messed up. That indicates, that the paste-plugin of tinymce suffers from the same issue in the 4.x stream and it has only been fixed in the 5.x stream.

(cherry picked from commit 095d16d0b1fb9a75dfed88c2e37c4fbfa19fcd38)