Fix: OXUIB-2014: Last row in calendar month view halfway missing
Root cause: Originally the height of the rows in the calendar's month
view were only calculated for Firefox and Internet Explorer. For this
reason the issue was less apparent in Firefox. Additionally, exact
calculation of the row height led to rounding errors when the calender is
rendering in Firefox and Chromium based browsers. The issue could only be
observed on months spanning 6 weeks, because there the calculated row
height equals 100% / 6 = 16.66666666...%.

Solution: Apply calculation of row heights to all browsers and round down
the results.